+++ apps/storybook2/src/stories/package/trading-rewards/overview.stories.tsx
@@ -20,12 +20,12 @@ const meta = {
       // const networkId = localStorage.getItem("preview-orderly-networkId");
       // const networkId = "mainnet";
       const networkId = "testnet";
-      const configStore = new CustomConfigStore({ networkId, brokerId: "woofi_pro", env: "qa" });
+      const configStore = new CustomConfigStore({ networkId, brokerId: "woofi_pro", env: "qa", brokerName:'WOOFi Pro' });
       return (
         <ConnectorProvider>
           <OrderlyApp
-            brokerId={"orderly"}
-            brokerName={"Orderly"}
+            // brokerId={"orderly"}
+            // brokerName={"Orderly"}
             networkId={networkId}
             configStore={configStore}
           >