+++ packages/markets/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @orderly.network/markets
 
+## 1.0.1-sdk-v2.79
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/hooks@1.5.10-sdk-v2.79
+  - @orderly.network/types@1.5.10-sdk-v2.79
+  - @orderly.network/utils@1.5.10-sdk-v2.79
+  - @orderly.network/ui@1.0.1-sdk-v2.79
+
 ## 1.0.1-sdk-v2.78
 
 ### Patch Changes