+++ packages/ui-transfer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @orderly.network/ui-transfer
 
+## 1.0.1-sdk-v2.79
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/ui-chain-selector@1.0.1-sdk-v2.79
+  - @orderly.network/ui-connector@1.0.1-sdk-v2.79
+  - @orderly.network/hooks@1.5.10-sdk-v2.79
+  - @orderly.network/types@1.5.10-sdk-v2.79
+  - @orderly.network/utils@1.5.10-sdk-v2.79
+  - @orderly.network/react-app@1.0.1-sdk-v2.79
+  - @orderly.network/ui@1.0.1-sdk-v2.79
+
 ## 1.0.1-sdk-v2.78
 
 ### Patch Changes